[FIX] pdf 관련 파일들 따로 격리
